refactor: model select for side panel switcher
