+++ client/src/hooks/Nav/useSideNavLinks.ts
@@ -1,31 +1,41 @@
 import { useMemo } from 'react';
 import { ArrowRightToLine, Settings2 } from 'lucide-react';
+import { EModelEndpoint } from 'librechat-data-provider';
 import type { TConfig } from 'librechat-data-provider';
 import type { NavLink } from '~/common';
 import PanelSwitch from '~/components/SidePanel/Builder/PanelSwitch';
-import FilesPanel from '~/components/SidePanel/Files/Panel';
 import Parameters from '~/components/SidePanel/Parameters/Panel';
+import FilesPanel from '~/components/SidePanel/Files/Panel';
 import { Blocks, AttachmentIcon } from '~/components/svg';
 
 export default function useSideNavLinks({
   hidePanel,
   assistants,
   keyProvided,
+  endpoint,
 }: {
   hidePanel: () => void;
   assistants?: TConfig | null;
   keyProvided: boolean;
+  endpoint?: EModelEndpoint | null;
 }) {
   const Links = useMemo(() => {
     const links: NavLink[] = [];
-    links.push({
-      title: 'com_sidepanel_parameters',
-      label: '',
-      icon: Settings2,
-      id: 'parameters',
-      Component: Parameters,
-    });
-    if (assistants && assistants.disableBuilder !== true && keyProvided) {
+    if (endpoint !== EModelEndpoint.assistants) {
+      links.push({
+        title: 'com_sidepanel_parameters',
+        label: '',
+        icon: Settings2,
+        id: 'parameters',
+        Component: Parameters,
+      });
+    }
+    if (
+      endpoint === EModelEndpoint.assistants &&
+      assistants &&
+      assistants.disableBuilder !== true &&
+      keyProvided
+    ) {
       links.push({
         title: 'com_sidepanel_assistant_builder',
         label: '',
@@ -52,7 +62,7 @@ export default function useSideNavLinks({
     });
 
     return links;
-  }, [assistants, keyProvided, hidePanel]);
+  }, [assistants, keyProvided, hidePanel, endpoint]);
 
   return Links;
 }
